# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for KubeNDT (Kubernetes Network Digital Twin).",
        "title": "KubeNDT API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/kube/config": {
            "get": {
                "description": "Returns information about the currently loaded kubeconfig (contexts, current context, server URLs).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kubeconfig"
                ],
                "summary": "Get kubeconfig info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.KubeConfigInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Loads kubeconfig from an uploaded file.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kubeconfig"
                ],
                "summary": "Load kubeconfig file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Kubeconfig file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.KubeConfigInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "descriptive error message"
                }
            }
        },
...
        "types.KubeConfigInfo": {
            "type": "object",
            "properties": {
                "configured": {
                    "type": "boolean",
                    "example": true
                },
                "context_cluster_ids": {
                    "description": "ContextClusterIDs maps each context name to its cluster's canonical ID\n(the kube-system namespace UID). Contexts whose cluster is currently\nunreachable are omitted.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "contexts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"my-cluster\"",
                        "\"staging\"]"
                    ]
                },
                "current_context": {
                    "type": "string",
                    "example": "my-cluster"
                },
                "path": {
                    "type": "string",
                    "example": "/home/user/.kube/config"
                }
            }
        },
...
    }
}